requirements.txt

flask
transformers
torch
torchvision
pillow
googletrans==4.0.0-rc1
